api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
paddleocr==2.7.0.3
paddlepaddle==2.5.2
transformers==4.38.2
torch==2.2.1
Pillow==10.2.0
numpy==1.24.3
opencv-python-headless==4.9.0.80
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
huggingface-hub==0.21.4